moved page to restaurants folder
